Update mygruvbox colorscheme
Mainly focused on the gui colors since I use them in the terminal as
well. Only set the expected background and foreground color on hi-Normal
and let the other entries fall back on either explicitly the bg/fg
color, or be transparent to draw properly on, e.g., float windows.